backend-django/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
crontab==0.23.0
cryptography==36.0.1
Django==3.2.6
django-cors-headers==3.10.0
django-crontab==0.7.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
et-xmlfile==1.1.0
h5py==3.7.0
idna==3.3
joblib==1.1.0
numpy==1.22.1
openpyxl==3.0.9
pandas==1.4.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
urllib3==1.26.9
yarg==0.1.9
